\name{ make.axion.map.to.h5.dh }
\alias{ make.axion.map.to.h5.dh }
%- Also NEED an '\alias' for EACH other topic documented here.

\title{
make.axion.map.to.h5.dh
}

\description{
A function that calls for user to input experiment log .csv file and .mapTimestamp files
(text files created in neuroExplorer with spike trains) and prints out .h5 files in a directory in the same directory as the folder containing the .mapTimestamp files.  A wrapper function for axion.map.to.h5.dh function.  
}

\usage{
make.axion.map.to.h5.dh( )
}

%- maybe also 'usage' for other objects documented here.

\arguments{

}


\examples{

\dontrun{ make.axion.map.to.h5.dh(  )  }

}


